refactor: migrate clash client info retrieval to MihomoManager
